+++ src/findPupils.hpp
@@ -22,11 +22,11 @@ namespace Detector {
 			if (file) fclose(file);
 		}
 		
-		PointPair find( cv::Mat faceROI, RectPair eyes, cv::Point2f offset );
-		PointPair findCorners( cv::Mat faceROI, RectPair cornerRegion, cv::Point2f offset );
+		PointPair find( cv::Mat faceROI, RectPair eyes, cv::Point2i offset );
+		PointPair findCorners( cv::Mat faceROI, RectPair cornerRegion, cv::Point2i offset );
 		
 	private:
-		cv::Point2f findPupil( cv::Mat &faceImage, cv::Rect2f &eyeRegion, bool isLeftEye );
+		cv::Point2f findPupil( cv::Mat &faceImage, cv::Rect2i &eyeRegion, bool isLeftEye );
 	};
 }
